+++ tests/datastore.py
@@ -196,6 +196,20 @@ TRAINTUPLE = {
                 "download": {"public": True, "authorized_ids": ["MyOrg1MSP"]},
             },
             "transient": False,
+            "value": {
+                "key": "6f0ee20a328044fb89e70ee5d219fa0c",
+                "compute_task_key": "30c283be-d385-424e-94a6-4d8538275260",
+                "address": {
+                    "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
+                    "storage_address": "http://backend-org-1-substra-backend-server.org-1:8000/model/5f0ee20a328044fb89e70ee5d219fa0b/address/",  # noqa: E501
+                },
+                "permissions": {
+                    "process": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+                    "download": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+                },
+                "owner": "MyOrg1MSP",
+                "creation_date": "2021-08-24T13:36:07.393646367Z",
+            },
         },
     },
     "train": {
@@ -209,7 +223,6 @@ TRAINTUPLE = {
         "models": [
             {
                 "key": "6f0ee20a328044fb89e70ee5d219fa0c",
-                "category": "MODEL_SIMPLE",
                 "compute_task_key": "30c283be-d385-424e-94a6-4d8538275260",
                 "address": {
                     "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
@@ -270,13 +283,26 @@ AGGREGATETUPLE = {
                 "download": {"public": True, "authorized_ids": ["MyOrg1MSP"]},
             },
             "transient": False,
+            "value": {
+                "key": "5f0ee20a328044fb89e70ee5d219fa0b",
+                "compute_task_key": "06207faf-1785-4fa9-4220-99a50dcfe064",
+                "address": {
+                    "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
+                    "storage_address": "http://backend-org-1-substra-backend-server.org-1:8000/model/5f0ee20a328044fb89e70ee5d219fa0b/address/",  # noqa: E501
+                },
+                "permissions": {
+                    "process": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+                    "download": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+                },
+                "owner": "MyOrg1MSP",
+                "creation_date": "2021-08-24T13:36:07.393646367Z",
+            },
         },
     },
     "aggregate": {
         "models": [
             {
                 "key": "5f0ee20a328044fb89e70ee5d219fa0b",
-                "category": "MODEL_SIMPLE",
                 "compute_task_key": "06207faf-1785-4fa9-4220-99a50dcfe064",
                 "address": {
                     "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
@@ -348,6 +374,20 @@ COMPOSITE_TRAINTUPLE = {
                 "download": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
             },
             "transient": False,
+            "value": {
+                "key": "dc9d114d1ed54936b57dd7fef1c0cbf4",
+                "compute_task_key": "aa09180a-fec6-46a5-a1a7-58c971b39217",
+                "address": {
+                    "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
+                    "storage_address": "http://backend-org-1-substra-backend-server.org-1:8000/model/ac4a1a82ced4419b83f4bba78c024b94/address/",  # noqa: E501
+                },
+                "permissions": {
+                    "process": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+                    "download": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+                },
+                "owner": "MyOrg1MSP",
+                "creation_date": "2021-08-24T13:36:07.393646367Z",
+            },
         },
         "local": {
             "permissions": {
@@ -355,6 +395,20 @@ COMPOSITE_TRAINTUPLE = {
                 "download": {"public": False, "authorized_ids": ["MyOrg2MSP"]},
             },
             "transient": False,
+            "value": {
+                "key": "ac4a1a82ced4419b83f4bba78c024b94",
+                "compute_task_key": "aa09180a-fec6-46a5-a1a7-58c971b39217",
+                "address": {
+                    "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
+                    "storage_address": "http://backend-org-1-substra-backend-server.org-1:8000/model/ac4a1a82ced4419b83f4bba78c024b94/address/",  # noqa: E501
+                },
+                "permissions": {
+                    "process": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+                    "download": {"public": False, "authorized_ids": ["MyOrg1MSP", "MyOrg2MSP"]},
+                },
+                "owner": "MyOrg1MSP",
+                "creation_date": "2021-08-24T13:36:07.393646367Z",
+            },
         },
     },
     "composite": {
@@ -363,7 +417,6 @@ COMPOSITE_TRAINTUPLE = {
         "models": [
             {
                 "key": "ac4a1a82ced4419b83f4bba78c024b94",
-                "category": "MODEL_HEAD",
                 "compute_task_key": "aa09180a-fec6-46a5-a1a7-58c971b39217",
                 "address": {
                     "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
@@ -378,7 +431,6 @@ COMPOSITE_TRAINTUPLE = {
             },
             {
                 "key": "dc9d114d1ed54936b57dd7fef1c0cbf4",
-                "category": "MODEL_SIMPLE",
                 "compute_task_key": "aa09180a-fec6-46a5-a1a7-58c971b39217",
                 "address": {
                     "checksum": "40483cd8b99ea7fbd3b73020997ea07547771993a6a3fa56fa2a8e9d7860529e",
@@ -735,7 +787,6 @@ MODEL = {
         "storage_address": "http://substra-backend.org-1.com/model/d7d88d74-d42f-43f4-affa-d0cdb99873fc/file/",
         "checksum": "b7881c547f7b80f01200fd2c41546010a15c66e2db46722dbbffa428a0bd0a26",
     },
-    "category": "MODEL_SIMPLE",
     "compute_task_key": "b2b37c0f-de0a-4a61-9d00-58079efa9b33",
     "creation_date": "2022-03-01T11:22:35.791650Z",
     "key": "d7d88d74-d42f-43f4-affa-d0cdb99873fc",